[DNM] exynos7420: resolve compilation errors

+++ mobicore/daemon/Common/CSemaphore.cpp
@@ -67,7 +67,7 @@ void CSemaphore::wait()
 }
 
 //------------------------------------------------------------------------------
-bool CSemaphore::wait(int sec)
+bool CSemaphore::wait(long int sec)
 {
     int rc = 0;
     struct timespec tm;
